bot/requirements.txt
faiss-cpu
fastapi
gunicorn
httpx
numpy
openai
psycopg2-binary
pydantic
pytest
python-dotenv
sqlalchemy
tiktoken
uvicorn